cmd directory moved to project root
