+++ cmd/display.go
